# docs/swagger.yaml
schemes:
  - http
  - https
swagger: '2.0'
info:
  description: REST API for the LMS platform covering auth, courses, enrollments, assessments, payments, and more.
  title: LMS Backend API
  contact: {}
  version: 1.0.0
host: localhost:8080
basePath: /
paths:
...
  /v1/admin/notifications/broadcast:
    post:
      security:
        - BearerAuth: []
      description: Sends a broadcast notification to all users or a specific role
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - notifications
      summary: Send broadcast notification
      parameters:
        - description: Broadcast data
          name: body
          in: body
          required: true
          schema:
            type: object
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/notifications.BroadcastResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/admin/notifications/templates:
    get:
      security:
        - BearerAuth: []
      description: Returns configurable notification templates
      produces:
        - application/json
      tags:
        - notifications
      summary: List notification templates
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/notifications.TemplateResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/admin/notifications/templates/{templateId}:
    patch:
      security:
        - BearerAuth: []
      description: Updates a notification template's subject and body
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - notifications
      summary: Update notification template
      parameters:
        - type: string
          description: Template ID
          name: templateId
          in: path
          required: true
        - description: Template update data
          name: body
          in: body
          required: true
          schema:
            type: object
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/notifications.TemplateResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
  /v1/notifications:
    get:
      security:
        - BearerAuth: []
      description: Returns a paginated list of notifications for the authenticated user
      produces:
        - application/json
      tags:
        - notifications
      summary: List notifications
      parameters:
        - type: integer
          default: 1
          description: Page number
          name: page
          in: query
        - type: integer
          default: 20
          description: Items per page
          name: limit
          in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/notifications.NotificationListResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/notifications/read-all:
    patch:
      security:
        - BearerAuth: []
      description: Marks all notifications as read for the authenticated user
      produces:
        - application/json
      tags:
        - notifications
      summary: Mark all notifications as read
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/notifications.MarkAllReadResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/notifications/{notificationId}/read:
    patch:
      security:
        - BearerAuth: []
      description: Marks a single notification as read for the authenticated user
      produces:
        - application/json
      tags:
        - notifications
      summary: Mark notification as read
      parameters:
        - type: string
          description: Notification ID
          name: notificationId
          in: path
          required: true
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
definitions:
...
  handlers.ErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          message:
            type: string
  handlers.ValidationDetail:
    type: object
    properties:
      field:
        type: string
      message:
        type: string
  handlers.ValidationErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          details:
            type: array
            items:
              $ref: '#/definitions/handlers.ValidationDetail'
          message:
            type: string
...
  notifications.BroadcastResponse:
    type: object
    properties:
      recipient_count:
        type: integer
  notifications.MarkAllReadResponse:
    type: object
    properties:
      marked_count:
        type: integer
  notifications.NotificationListResponse:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/notifications.NotificationResponse'
      meta:
        type: object
        additionalProperties: {}
      unread_count:
        type: integer
  notifications.NotificationResponse:
    type: object
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      title:
        type: string
      type:
        type: string
  notifications.TemplateResponse:
    type: object
    properties:
      allowed_variables:
        type: array
        items:
          type: string
      body_template:
        type: string
      channel:
        type: string
        enum:
          - in_app
          - email
          - both
      id:
        type: string
      subject_template:
        type: string
      type:
        type: string
      updated_at:
        type: string
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header